Extract user validation utils to own module in users folder
