+++ src/handlers.utils.js
@@ -1,7 +1,5 @@
 const R = require('ramda');
-const SlackTemplate = require('claudia-bot-builder').slackTemplate;
 const HandlerError = require('./HandlerError');
-const { formatUser } = require('./formatting');
 
 const chain = (...fs) => deps =>
     R.pipe(
@@ -17,39 +15,6 @@ const concurrent = (...fs) => deps => message =>
         vs => Promise.all(vs)
             .then(R.reduce((acc, v) => ({ ...acc, ...v }), {})),
     )(fs);
-
-const validateUser = ({ userRepository }) => async message => {
-    const { sender } = message;
-    const user = await userRepository.getUser(sender);
-    return user
-        ? Promise.resolve({ ...message, user })
-        : Promise.reject(new HandlerError(
-            new SlackTemplate('I don\'t know who you are. :crying_cat_face:')
-                .replaceOriginal(false)
-                .get(),
-        ));
-};
-
-const validateNoUser = ({ userRepository }) => async message => {
-    const { sender } = message;
-    const user = await userRepository.getUser(sender);
-    return user
-        ? Promise.reject(new HandlerError(
-            new SlackTemplate(`You are already logged in as ${formatUser(user)}. :angry:`)
-                .replaceOriginal(false)
-                .get(),
-        ))
-        : Promise.resolve(message);
-};
-
-const tryGetUser = ({ userRepository }) => async message => {
-    const { sender } = message;
-    const user = await userRepository.getUser(sender);
-    return Promise.resolve(user
-        ? { ...message, user }
-        : message,
-    );
-};
 
 const validateCallbackValue = (actionName, valuePropName = 'callbackValue') => () => message => {
     const { originalRequest } = message;
@@ -84,9 +49,6 @@ const parseCallbackValue = (actionName, { actions }) =>
 module.exports = {
     chain,
     concurrent,
-    validateUser,
-    validateNoUser,
-    tryGetUser,
     validateCallbackValue,
     tryGetCallbackValue,
     parseCallbackValue,
