+++ src/users/formatting.js
@@ -1,0 +1,6 @@
+
+const formatUser = ({ challongeUsername, challongeEmailHash }) => `*${challongeUsername}* (${challongeEmailHash ? 'verified' : 'unverified'})`;
+
+module.exports = {
+    formatUser,
+};
