+++ src/users/index.js
@@ -1,4 +1,5 @@
 
 module.exports = {
     handlers: require('./handlers'),
+    validation: require('./validation'),
 };
